FastHistograms

Documentation for FastHistograms.

FastHistograms.FastHistograms

FastHistograms.SingleThreadFixedWidth2DHistogram

FastHistograms.SingleThreadFixedWidth2DHistogram

FastHistograms.SingleThreadFixedWidth2DHistogram

FastHistograms.bin_type

FastHistograms.calc_hist!

FastHistograms.counts

FastHistograms.zero!

FastHistograms.FastHistograms — Module

FastHistograms declares and implements a minimal histogram interface with a focus on speed.

using FastHistograms, Random

h = SingleThreadFixedWidth2DHistogram()

img1 = zeros(10, 10)
img1[:, 6:end] .= 0xff

img2 = zeros(10, 10)
img2[6:end, :] .= 0xff

# Compute the histogram bin counts
calc_hist!(h, img1, img2)

@show counts(h) # Get the bin counts

FastHistograms.SingleThreadFixedWidth2DHistogram — Type

A 2D histogram with fixed-width bins. The histogram computation is single-threaded by design and is therefore best suited to small data.

FastHistograms.SingleThreadFixedWidth2DHistogram — Type

SingleThreadFixedWidth2DHistogram(bins = 0x00:UInt8(16):0xff)

A 2D histogram with fixed-width bins. The histogram computation is single-threaded by design and is therefore best suited to small data.

The bins must be given in strictly increasing order (e.g., [0, 64, 128, 192]).

FastHistograms.SingleThreadFixedWidth2DHistogram — Method

SingleThreadFixedWidth2DHistogram(
    first_bin::BinType,
    last_bin::BinType,
    nbins::Int,
) where {BinType<:Integer}

A 2D histogram with fixed-width bins. The histogram computation is single-threaded by design and is therefore best suited to small data.

The first_bin and last_bin are the values of the lowest and highest bins, respectively. nbins is the total number of bins.

FastHistograms.bin_type — Function

Returns the type of the bin used in the histogram. This function takes one argument: the histogram structure.

This function is not required to be thread-safe.

FastHistograms.calc_hist! — Function

Computes the histogram of the given data. This function takes two or three arguments. The first argument must be the histogram structure (a subtype of AbstractHistogram). The second and optional third argument(s) must be the data to operate on. The type of the histogram determines whether one or two additional arguments are needed. For example, a 2D histogram requires two additional arguments. A typical 1D histogram requires only one.

Implementations of this function are free to implement any optimizations they wish as long as the histogram is correct once this function exits. This function is not required to be thread-safe. Implementations are encouraged to create descriptive names for their subtypes of AbstractHistogram that describe which optimizations are implemented. For example, SingleThreadFixedWidth2DHistogram does not implement thread-level parallelization by design.

FastHistograms.counts — Function

Returns the count in each bin of the histogram. This function takes one argument: the histogram structure.

This function is not required to be thread-safe.

FastHistograms.zero! — Function

Sets the count in each bin to zero. This function takes one argument: the histogram structure.

This function is not required to be thread-safe.